- Moved admin modules to (modules) folder grouping.
