+++ src/app/admin/(modules)/daydreams/[daydreamId]/loading.tsx
@@ -1,4 +1,4 @@
-import DaydreamFormLoading from "@/app/admin/daydreams/_components/Loading/DaydreamFormLoading";
+import DaydreamFormLoading from "@/app/admin/(modules)/daydreams/_components/Loading/DaydreamFormLoading";
 
 const EditDaydreamLoading = () => (
   <DaydreamFormLoading submitButtonText="Update Dream" title="Edit Dream" />
