+++ src/app/admin/(modules)/daydreams/_components/DaydreamForm.tsx
@@ -34,8 +34,8 @@ import { any, number, object, string } from "zod";
 import { FileAPIDataStructure } from "@/features/files/types";
 import { FaRegImages } from "react-icons/fa6";
 import { PiXBold } from "react-icons/pi";
-import GalleryDialog from "@/app/admin/daydreams/_components/GalleryDialog";
-import SupabaseImage from "@/app/admin/daydreams/_components/SupabaseImage";
+import GalleryDialog from "@/app/admin/(modules)/daydreams/_components/GalleryDialog";
+import SupabaseImage from "@/app/admin/(modules)/daydreams/_components/SupabaseImage";
 import classNames from "classnames";
 
 export const DreamSchema = object({
