+++ src/app/admin/(modules)/daydreams/_components/SupabaseImage.tsx
