+++ src/app/admin/(modules)/daydreams/create/_components/CreateDaydreamWrapper.tsx
@@ -4,7 +4,7 @@ import { DreamFormParams } from "@/features/daydreams/types";
 import { DEFAULT_FORM_VALUES } from "@/features/daydreams/data";
 import { useCallback } from "react";
 import { useStoreDaydreamMutation } from "@/features/daydreams/api";
-import DaydreamForm from "@/app/admin/daydreams/_components/DaydreamForm";
+import DaydreamForm from "@/app/admin/(modules)/daydreams/_components/DaydreamForm";
 import toast from "react-hot-toast";
 
 const CreateDaydreamWrapper = () => {
