+++ src/app/admin/(modules)/daydreams/create/loading.tsx
@@ -1,4 +1,4 @@
-import DaydreamFormLoading from "@/app/admin/daydreams/_components/Loading/DaydreamFormLoading";
+import DaydreamFormLoading from "@/app/admin/(modules)/daydreams/_components/Loading/DaydreamFormLoading";
 
 const CreateDaydreamLoading = () => (
   <DaydreamFormLoading submitButtonText="Create" title="Create Daydream" />
