+++ src/app/admin/(modules)/daydreams/layout.tsx
@@ -1,9 +1,9 @@
 import { Fragment, Suspense } from "react";
-import DaydreamTableFilterLoading from "@/app/admin/daydreams/_components/Loading/DaydreamTableFilterLoading";
-import DaydreamTableFilters from "@/app/admin/daydreams/_components/DaydreamTableFilters";
-import DaydreamTable from "@/app/admin/daydreams/_components/DaydreamTable";
+import DaydreamTableFilterLoading from "@/app/admin/(modules)/daydreams/_components/Loading/DaydreamTableFilterLoading";
+import DaydreamTableFilters from "@/app/admin/(modules)/daydreams/_components/DaydreamTableFilters";
+import DaydreamTable from "@/app/admin/(modules)/daydreams/_components/DaydreamTable";
 import { BaseComponent } from "@/types";
-import DaydreamTableLoading from "@/app/admin/daydreams/_components/Loading/DaydreamTableLoading";
+import DaydreamTableLoading from "@/app/admin/(modules)/daydreams/_components/Loading/DaydreamTableLoading";
 
 const AdminDaydreamIndexLayout = ({
   children,
