+++ src/app/admin/(modules)/gallery/create/page.tsx
@@ -1,5 +1,5 @@
 import { Fragment } from "react";
-import { BreadcrumbDataSetter } from "@/app/admin/_components/Breadcrumbs";
+import { BreadcrumbDataSetter } from "src/app/admin/(modules)/_components/Breadcrumbs";
 import GalleryCreateWrapper from "@/app/admin/(modules)/gallery/create/_components/GalleryCreateWrapper";
 import { Metadata } from "next";
