+++ src/app/admin/(modules)/projects/_components/ProjectCard.tsx
@@ -11,7 +11,7 @@ import { Tables } from "@/types";
 import { CSS } from "@dnd-kit/utilities";
 import { Fragment, useState } from "react";
 import classNames from "classnames";
-import SupabaseImage from "@/app/admin/daydreams/_components/SupabaseImage";
+import SupabaseImage from "@/app/admin/(modules)/daydreams/_components/SupabaseImage";
 import ImagePreviewDialog from "@/app/admin/(modules)/projects/_components/ImagePreviewDialog";
 import toast from "react-hot-toast";
