+++ src/app/admin/(modules)/projects/page.tsx
@@ -1,5 +1,5 @@
 import { Fragment } from "react";
-import { BreadcrumbDataSetter } from "@/app/admin/_components/Breadcrumbs";
+import { BreadcrumbDataSetter } from "src/app/admin/(modules)/_components/Breadcrumbs";
 import { Metadata } from "next";
 import IndexProjectListWrapper from "@/app/admin/(modules)/projects/_components/IndexProjectListWrapper";
 import IndexProjectHeader from "@/app/admin/(modules)/projects/_components/IndexProjectHeader";
